Refactor: Removing birthDate field from Person entity and adding new attribute called age, ajusting validators and removing pagination

+++ Back-end/VaccinationCard/VaccinationCard.Application/Vaccines/Commands/DeleteVaccine/DeleteVaccineCommandHandler.cs
@@ -4,19 +4,27 @@ using System.Net;
 using VaccinationCard.Application.Interfaces.Repositories;
 using VaccinationCard.Domain.Entities;
 using VaccinationCard.Domain.Errors;
+using VaccinationCard.Domain.Interfaces.Repositories;
 
 namespace VaccinationCard.Application.Vaccines.Commands.DeleteVaccine
 {
-    internal class DeleteVaccineCommandHandler(IBaseRepository<Vaccine> vaccineRepository) : IRequestHandler<DeleteVaccineCommand, Result>
+    internal class DeleteVaccineCommandHandler(IBaseRepository<Vaccine> vaccineRepository, IVaccinationRepository vaccinationRepository) : IRequestHandler<DeleteVaccineCommand, Result>
     {
 
         private readonly IBaseRepository<Vaccine> _vaccineRepository = vaccineRepository;
+
+        private readonly IVaccinationRepository _vaccinationRepository = vaccinationRepository;
 
         public async Task<Result> Handle(DeleteVaccineCommand request, CancellationToken cancellationToken)
         {
             var vaccine = await _vaccineRepository.GetByIdAsync(request.VaccineId);
 
             if (vaccine == null) return Result.Failure(VaccineErrors.NotFound, HttpStatusCode.NotFound);
+
+            if(await _vaccinationRepository.IsVaccineBeingUsed(request.VaccineId))
+            {
+                return Result.Failure(VaccineErrors.VaccineIsBeingUsed, HttpStatusCode.Conflict);
+            }
 
             await _vaccineRepository.DeleteAsync(request.VaccineId);
